requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
loguru
prometheus-fastapi-instrumentator
langchain
langgraph
chromadb
sentence-transformers
pandas
numpy
scikit-learn
pymupdf
openai
tiktoken
pinecone-client
pytest
httpx
